dont delete the postmeta if video or caption url

+++ classes/class-cove-asset-metaboxes.php
@@ -426,7 +426,9 @@ class COVE_Asset_Metaboxes {
   			}
 		
   		 	if ( ${$f} == '' ) { 
-  		 		delete_post_meta( $post_id , $f , get_post_meta( $post_id , $f , true ) );
+          if ( ($f != '_coveam_video_url') && ($f != '_coveam_video_caption') ) {
+  		 		  delete_post_meta( $post_id , $f , get_post_meta( $post_id , $f , true ) );
+          }
   		 	} else {
   		 		update_post_meta( $post_id , $f , ${$f} );
   		 	}
